requirements.txt
Pillow
MIDIUtil
numpy
django-widget-tweaks
scipy